feat: test unitário remoção usuario aluguel aberto

+++ bike/src/app.ts
@@ -10,7 +10,7 @@ import { DuplicateUserError } from "./errors/duplicate-user-error";
 import { RentRepo } from "./ports/rent-repo";
 import { UserRepo } from "./ports/user-repo";
 import { BikeRepo } from "./ports/bike-repo";
-import { RentNotFoundError } from "./errors/rent-not-found-error";
+import { RentOpenError } from "./errors/rent-open-error";
 
 export class App {
     crypt: Crypt = new Crypt()
@@ -19,7 +19,7 @@ export class App {
         readonly userRepo: UserRepo,
         readonly bikeRepo: BikeRepo,
         readonly rentRepo: RentRepo
-    ) {}
+    ) { }
 
     async findUser(email: string): Promise<User> {
         const user = await this.userRepo.find(email)
@@ -27,9 +27,13 @@ export class App {
         return user
     }
 
+    async findOpenRentsFor(userEmail: string): Promise<boolean> {
+        return this.rentRepo.findOpenRentsFor(userEmail);
+    }
+
     async registerUser(user: User): Promise<string> {
         if (await this.userRepo.find(user.email)) {
-          throw new DuplicateUserError()
+            throw new DuplicateUserError()
         }
         const encryptedPassword = await this.crypt.encrypt(user.password)
         user.password = encryptedPassword
@@ -47,9 +51,12 @@ export class App {
 
     async removeUser(email: string): Promise<void> {
         await this.findUser(email)
-        await this.userRepo.remove(email)
+        if (await this.findOpenRentsFor(email))
+            throw new RentOpenError()
+        else
+            await this.userRepo.remove(email)
     }
-    
+
     async rentBike(bikeId: string, userEmail: string): Promise<string> {
         const bike = await this.findBike(bikeId)
         if (!bike.available) {
@@ -65,7 +72,7 @@ export class App {
     async returnBike(bikeId: string, userEmail: string): Promise<number> {
         const now = new Date()
         const rent = await this.rentRepo.findOpen(bikeId, userEmail)
-        if (!rent) throw new RentNotFoundError();
+        if (!rent) throw new Error('Rent not found.')
         rent.end = now
         await this.rentRepo.update(rent.id, rent)
         rent.bike.available = true
@@ -97,7 +104,7 @@ export class App {
 }
 
 function diffHours(dt2: Date, dt1: Date) {
-  var diff = (dt2.getTime() - dt1.getTime()) / 1000;
-  diff /= (60 * 60);
-  return Math.abs(diff);
+    var diff = (dt2.getTime() - dt1.getTime()) / 1000;
+    diff /= (60 * 60);
+    return Math.abs(diff);
 }
